Move `inspector` to its own file.
This commit prepares for the `data_converter`, `Reader` and `Writer`
to depend on `DataType`. This would cause the includes to become cyclic
since the `DataType` requires the inspector.

Note that this commit only moves code from one file to the next without
any other changes.
